+++ include/aws/http/connection_manager.h
@@ -24,21 +24,6 @@ typedef void(aws_http_connection_manager_on_connection_setup_fn)(
 
 typedef void(aws_http_connection_manager_shutdown_complete_fn)(void *user_data);
 
-enum aws_connection_manager_proxy_env_var_type {
-    /**
-     * Default.
-     * Disable reading from environment variable for proxy.
-     */
-    AWS_CMPEV_DISABLE = 0,
-    /**
-     * Enable get proxy URL from environment variable, when the manual proxy options of connection manager is not set.
-     * env HTTPS_PROXY/https_proxy will be checked when the main connection use tls.
-     * env HTTP_PROXY/http_proxy will be checked when the main connection NOT use tls.
-     * The lower case version has precedence.
-     */
-    AWS_CMPEV_ENABLE,
-};
-
 /*
  * Connection manager configuration struct.
  *
@@ -58,27 +43,13 @@ struct aws_http_connection_manager_options {
     uint16_t port;
     /* Proxy configuration for http connection */
     const struct aws_http_proxy_options *proxy_options;
+
     /*
      * Optional.
      * Configuration for using proxy from environment variable.
-     * Only works when proxy_options is not set. Zero out as default settings.
+     * Only works when proxy_options is not set.
      */
-    struct {
-        enum aws_connection_manager_proxy_env_var_type env_var_type;
-        /*
-         * Optional.
-         * If not set:
-         * If tls options are provided (for the main connection) use tunnel proxy type
-         * If tls options are not provided (for the main connection) use forward proxy type
-         */
-        enum aws_http_proxy_connection_type connection_type;
-        /*
-         * Optional.
-         * If not set, a default tls option will be created. when https used for Local to proxy connection.
-         * Must be distinct from the the tls_connection_options from aws_http_connection_manager_options
-         */
-        const struct aws_tls_connection_options *tls_options;
-    } proxy_env_var_settings;
+    const struct proxy_env_var_settings *proxy_ev_settings;
 
     /*
      * Maximum number of connections this manager is allowed to contain
