+++ source/connection.c
@@ -11,19 +11,27 @@
 
 #include <aws/http/private/proxy_impl.h>
 
+#include <aws/common/environment.h>
 #include <aws/common/hash_table.h>
 #include <aws/common/mutex.h>
 #include <aws/common/string.h>
+#include <aws/http/proxy.h>
 #include <aws/http/request_response.h>
 #include <aws/io/channel_bootstrap.h>
 #include <aws/io/logging.h>
 #include <aws/io/socket.h>
 #include <aws/io/tls_channel_handler.h>
+#include <aws/io/uri.h>
 
 #if _MSC_VER
 #    pragma warning(disable : 4204) /* non-constant aggregate initializer */
 #    pragma warning(disable : 4232) /* function pointer to dll symbol */
 #endif
+
+AWS_STATIC_STRING_FROM_LITERAL(s_http_proxy_env_var, "HTTP_PROXY");
+AWS_STATIC_STRING_FROM_LITERAL(s_http_proxy_env_var_low, "http_proxy");
+AWS_STATIC_STRING_FROM_LITERAL(s_https_proxy_env_var, "HTTPS_PROXY");
+AWS_STATIC_STRING_FROM_LITERAL(s_https_proxy_env_var_low, "https_proxy");
 
 static struct aws_http_connection_system_vtable s_default_system_vtable = {
     .new_socket_channel = aws_client_bootstrap_new_socket_channel,
@@ -885,6 +893,122 @@ static int s_validate_http_client_connection_options(const struct aws_http_clien
     return AWS_OP_SUCCESS;
 }
 
+static int s_proxy_uri_init_from_env_variable(
+    struct aws_allocator *allocator,
+    const struct aws_http_client_connection_options *options,
+    struct aws_uri *proxy_uri,
+    bool *found) {
+    struct aws_string *proxy_uri_string = NULL;
+    *found = false;
+    if (options->tls_options) {
+        if (aws_get_environment_value(allocator, s_https_proxy_env_var_low, &proxy_uri_string) == AWS_OP_SUCCESS &&
+            proxy_uri_string != NULL) {
+            AWS_LOGF_DEBUG(AWS_LS_HTTP_CONNECTION_MANAGER, "https_proxy environment found");
+        } else if (
+            aws_get_environment_value(allocator, s_https_proxy_env_var, &proxy_uri_string) == AWS_OP_SUCCESS &&
+            proxy_uri_string != NULL) {
+            AWS_LOGF_DEBUG(AWS_LS_HTTP_CONNECTION_MANAGER, "HTTPS_PROXY environment found");
+        } else {
+            return AWS_OP_SUCCESS;
+        }
+    } else {
+        if (aws_get_environment_value(allocator, s_http_proxy_env_var_low, &proxy_uri_string) == AWS_OP_SUCCESS &&
+            proxy_uri_string != NULL) {
+            AWS_LOGF_DEBUG(AWS_LS_HTTP_CONNECTION_MANAGER, "http_proxy environment found");
+        } else if (
+            aws_get_environment_value(allocator, s_http_proxy_env_var, &proxy_uri_string) == AWS_OP_SUCCESS &&
+            proxy_uri_string != NULL) {
+            AWS_LOGF_DEBUG(AWS_LS_HTTP_CONNECTION_MANAGER, "HTTP_PROXY environment found");
+        } else {
+            return AWS_OP_SUCCESS;
+        }
+    }
+    struct aws_byte_cursor proxy_uri_cursor = aws_byte_cursor_from_string(proxy_uri_string);
+    if (aws_uri_init_parse(proxy_uri, allocator, &proxy_uri_cursor)) {
+        AWS_LOGF_ERROR(AWS_LS_HTTP_CONNECTION_MANAGER, "Could not parse found proxy URI.");
+        aws_string_destroy(proxy_uri_string);
+        return AWS_OP_ERR;
+    }
+    *found = true;
+    aws_string_destroy(proxy_uri_string);
+    return AWS_OP_SUCCESS;
+}
+
+static int s_init_proxy_options_from_env_variable(
+    const struct aws_http_client_connection_options *options,
+    struct aws_http_proxy_options *proxy_options) {
+
+    struct aws_uri proxy_uri;
+    AWS_ZERO_STRUCT(proxy_uri);
+    bool found = false;
+    bool success = false;
+    if (s_proxy_uri_init_from_env_variable(options->allocator, options, &proxy_uri, &found)) {
+        /* Envrionment is set but failed to parse it */
+        goto done;
+    }
+    if (found) {
+        proxy_options->host = proxy_uri.host_name;
+        proxy_options->port = proxy_uri.port;
+        struct aws_tls_connection_options default_tls_connection_options;
+        AWS_ZERO_STRUCT(default_tls_connection_options);
+        proxy_options->connection_type = options->proxy_ev_settings->connection_type;
+        if (!proxy_options->connection_type) {
+            if (options->tls_options) {
+                /* Use Tunneling when main connection use TLS. */
+                proxy_options->connection_type = AWS_HPCT_HTTP_TUNNEL;
+            } else {
+                /* Use forwarding proxy when main connection use clear text. */
+                proxy_options->connection_type = AWS_HPCT_HTTP_FORWARD;
+            }
+        }
+        if (aws_byte_cursor_eq_ignore_case(&proxy_uri.scheme, &aws_http_scheme_https)) {
+            if (options->proxy_ev_settings->tls_options) {
+                proxy_options->tls_options = options->proxy_ev_settings->tls_options;
+            } else {
+                struct aws_tls_ctx *tls_ctx = NULL;
+                struct aws_tls_ctx_options tls_ctx_options;
+                AWS_ZERO_STRUCT(tls_ctx_options);
+                /* create a default tls options */
+                aws_tls_ctx_options_init_default_client(&tls_ctx_options, options->allocator);
+                /* turn off verify peer for proxy connection setup by default */
+                aws_tls_ctx_options_set_verify_peer(&tls_ctx_options, false);
+                tls_ctx = aws_tls_client_ctx_new(options->allocator, &tls_ctx_options);
+                aws_tls_ctx_options_clean_up(&tls_ctx_options);
+                if (!tls_ctx) {
+                    AWS_LOGF_ERROR(AWS_LS_HTTP_CONNECTION_MANAGER, "Failed to create default TLS context.");
+                    goto done;
+                }
+                aws_tls_connection_options_init_from_ctx(&default_tls_connection_options, tls_ctx);
+                /* tls options hold a ref to the ctx */
+                aws_tls_ctx_release(tls_ctx);
+                if (aws_tls_connection_options_set_server_name(
+                        &default_tls_connection_options, options->allocator, &proxy_uri.host_name)) {
+                    AWS_LOGF_ERROR(
+                        AWS_LS_HTTP_CONNECTION_MANAGER, "Failed set server name for TLS connection options.");
+                    goto done;
+                }
+                proxy_options->tls_options = &default_tls_connection_options;
+            }
+        }
+        /* Support basic authentication. */
+        if (proxy_uri.password.len) {
+            /* Has no empty password set */
+            struct aws_http_proxy_strategy_basic_auth_options config = {
+                .proxy_connection_type = proxy_options->connection_type,
+                .user_name = proxy_uri.user,
+                .password = proxy_uri.password,
+            };
+            struct aws_http_proxy_strategy *proxy_strategy =
+                aws_http_proxy_strategy_new_basic_auth(options->allocator, &config);
+            proxy_options->proxy_strategy = proxy_strategy;
+        }
+    }
+    success = true;
+done:
+    aws_uri_clean_up(&proxy_uri);
+    return success ? AWS_OP_SUCCESS : AWS_OP_ERR;
+}
+
 int aws_http_client_connect_internal(
     const struct aws_http_client_connection_options *orig_options,
     aws_http_proxy_request_transform_fn *proxy_request_transform) {
@@ -1018,7 +1142,19 @@ int aws_http_client_connect(const struct aws_http_client_connection_options *opt
     if (options->proxy_options != NULL) {
         return aws_http_client_connect_via_proxy(options);
     } else {
-        return aws_http_client_connect_internal(options, NULL);
+        if (!options->proxy_ev_settings || options->proxy_ev_settings->env_var_type != AWS_HPEV_ENABLE) {
+            return aws_http_client_connect_internal(options, NULL);
+        } else {
+            struct aws_http_proxy_options proxy_options;
+            AWS_ZERO_STRUCT(proxy_options);
+            if (s_init_proxy_options_from_env_variable(options, &proxy_options)) {
+                return AWS_OP_ERR;
+            }
+            struct aws_http_client_connection_options copied_options = *options;
+            copied_options.proxy_options = &proxy_options;
+            int return_code = aws_http_client_connect_via_proxy(&copied_options);
+            return return_code;
+        }
     }
 }
