+++ tests/CMakeLists.txt
@@ -459,8 +459,9 @@ add_net_test_case(test_connection_manager_proxy_setup_shutdown)
 add_net_test_case(test_connection_manager_idle_culling_single)
 add_net_test_case(test_connection_manager_idle_culling_many)
 add_net_test_case(test_connection_manager_idle_culling_mixture)
-add_net_test_case(test_connection_manager_read_http_proxy_ev)
-add_net_test_case(test_connection_manager_read_https_proxy_ev)
+# they should be moved to connection test?
+# add_net_test_case(test_connection_manager_read_http_proxy_ev)
+# add_net_test_case(test_connection_manager_read_https_proxy_ev)
 
 # tests where we establish real connections
 add_net_test_case(test_connection_manager_single_connection)
